moved some classes to the base of relatus
